requirements.txt
pandas
numpy
selenium
langchain-anthropic==0.1.23 ; python_version >= "3.11" and python_version < "3.12"
langchain-chroma==0.1.3 ; python_version >= "3.11" and python_version < "3.12"
langchain-cohere==0.1.9 ; python_version >= "3.11" and python_version < "3.12"
langchain-community==0.2.15 ; python_version >= "3.11" and python_version < "3.12"
langchain-core==0.2.37 ; python_version >= "3.11" and python_version < "3.12"
langchain-elasticsearch==0.2.2 ; python_version >= "3.11" and python_version < "3.12"
langchain-experimental==0.0.61 ; python_version >= "3.11" and python_version < "3.12"
langchain-google-genai==1.0.10 ; python_version >= "3.11" and python_version < "3.12"
langchain-huggingface==0.0.3 ; python_version >= "3.11" and python_version < "3.12"
langchain-milvus==0.1.4 ; python_version >= "3.11" and python_version < "3.12"
langchain-openai==0.1.23 ; python_version >= "3.11" and python_version < "3.12"
langchain-teddynote==0.0.28 ; python_version >= "3.11" and python_version < "3.12"
langchain-text-splitters==0.2.2 ; python_version >= "3.11" and python_version < "3.12"
langchain-upstage==0.1.8 ; python_version >= "3.11" and python_version < "3.12"
langchain==0.2.15 ; python_version >= "3.11" and python_version < "3.12"
langchainhub==0.1.21 ; python_version >= "3.11" and python_version < "3.12"
python-dotenv
google-search-results
streamlit